权限

+++ django/Kblog/manages.py
@@ -19,55 +19,54 @@ timeZone = pytz.timezone('Asia/Shanghai')
 manageThemeDir = 'manage/'
 
 def manage(request):
-	articleCount = Article.objects.count()
-	commentCount = Comment.objects.count()
-	articleList = Article.objects.all().order_by('-article_id')[:5]
-	commentList = Comment.objects.all().order_by('-comment_id')[:5]
-	context = {
-		'commentList'   : commentList,
-		'articleList'   : articleList,
-		'commentCount'  : commentCount,
-		'articleCount'  : articleCount
-	}
-	return render(request, manageThemeDir + 'manage.html', context)
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
+    else:
+		context = {}
+		articleCount = Article.objects.count()
+		commentCount = Comment.objects.count()
+		articleList = Article.objects.all().order_by('-article_id')[:5]
+		commentList = Comment.objects.all().order_by('-comment_id')[:5]
+
+		context = {
+			'commentList'   : commentList,
+			'articleList'   : articleList,
+			'commentCount'  : commentCount,
+			'articleCount'  : articleCount
+		}
+		return render(request, manageThemeDir + 'manage.html', context)
 
 
 #=======================================#
 
 
 def article(request):
-	context = {}
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
 
-	userInfo = request.session.get('uInfo', False)
-
-	if C.checkLoginAdmin(userInfo) == False:
-		return HttpResponseRedirect('/')
-
-	searchCondition = ''
-
-	keyword = cgi.escape(request.GET.get('word', ''))
-	if C.isset(keyword):
+    context = {}
+    searchCondition = ''
+    keyword = cgi.escape(request.GET.get('word', ''))
+    if C.isset(keyword):
 		searchCondition += " WHERE article.title LIKE '%%" + keyword + "%%' "
 
+    sql = 'SELECT article_id, title, username AS author, name AS category FROM "' + Meta.db_table + '_article" AS article '
+    sql += 'LEFT JOIN "' + Meta.db_table + '_relation" AS relation ON article.article_id=relation.aid '
+    sql += 'LEFT JOIN "' + Meta.db_table + '_category" AS category ON relation.cid=category.category_id '
+    sql += 'LEFT JOIN "' + Meta.db_table + '_user" AS usertab ON article.author=usertab.user_id '
+    sql += searchCondition + 'ORDER BY article_id DESC'
+    articleList = list(Article.objects.raw(sql))
 
-	sql = 'SELECT article_id, title, username AS author, name AS category FROM "' + Meta.db_table + '_article" AS article '
-	sql += 'LEFT JOIN "' + Meta.db_table + '_relation" AS relation ON article.article_id=relation.aid '
-	sql += 'LEFT JOIN "' + Meta.db_table + '_category" AS category ON relation.cid=category.category_id '
-	sql += 'LEFT JOIN "' + Meta.db_table + '_user" AS usertab ON article.author=usertab.user_id '
-	sql += searchCondition + 'ORDER BY article_id DESC'
+    paginator = Paginator(articleList, 10)
+    page = int(request.GET.get('page', 1))
 
-	articleList = list(Article.objects.raw(sql))
+    try:
+        pagebar = paginator.page(page)
+    except PageNotAnInteger:
+        pagebar = paginator.page(1)
+    except EmptyPage:
+        pagebar = paginator.page(paginator.num_pages)
 
-	paginator = Paginator(articleList, 10)
-	page = int(request.GET.get('page', 1))
-
-	try:
-		pagebar = paginator.page(page)
-	except PageNotAnInteger:
-		pagebar = paginator.page(1)
-	except EmptyPage:
-		pagebar = paginator.page(paginator.num_pages)
-	
 	context = {
 	    'pagebar' : pagebar
 	}
@@ -76,6 +75,9 @@ def article(request):
 
 
 def article_write(request):
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
+
 	userInfo  = request.session.get('uInfo', False)
 	url       = C.getProtocol(request) + request.get_host()
 
@@ -113,8 +115,10 @@ def article_write(request):
 
 
 def article_edit(request):
-	context = {}
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
 
+	context = {}
 	url       = request.get_host()
 	userInfo  = request.session.get('uInfo', False)
 	aid       = cgi.escape(request.GET.get('aid', 0))
@@ -164,6 +168,8 @@ def article_edit(request):
 
 
 def article_delete(request):
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
 	idList = request.REQUEST.getlist('idlist')
 	if len(idList) > 0:
 		Article.objects.filter(article_id__in=idList).delete()
@@ -176,6 +182,9 @@ def article_delete(request):
 
 
 def file_upload(request):
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponse('No signin')
+
 	responseData = {}
 	
 	url = C.getProtocol(request) + request.get_host()
@@ -230,13 +239,16 @@ def file_upload(request):
 		user_id        = int(userId),
 		original_name  = reqfile.name,
 		size           = int(fileSize),
-		savepath       = time.strftime('%Y/%m/',time.localtime(time.time()))
+		savepath       = time.strftime('%Y/%m/', time.localtime(time.time()))
 	)
 
 	return HttpResponse(json.dumps(responseData), content_type='application/json')
 
 
 def file_delete(request):
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
+
 	responseData = {}
 	responseData['status'] = 'faild'
 
@@ -253,6 +265,9 @@ def file_delete(request):
 
 
 def comment(request):
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
+
 	context       = {}
 	cid           = int(request.GET.get('cid', 0))
 	status        = cgi.escape(request.GET.get('status', 'pass'))
@@ -310,6 +325,8 @@ def comment(request):
 	return render(request, manageThemeDir + 'comment.html', context)
 
 def comment_delete(request):
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
 	idList = request.REQUEST.getlist('idlist')
 	if len(idList) > 0:
 		Comment.objects.filter(comment_id__in=idList).delete()
@@ -321,6 +338,9 @@ def comment_delete(request):
 
 
 def category(request):
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
+
 	context = {
 		'categoryList' : C.getCategoryList('<i class="uk-icon-minus"></i>')
 	}
@@ -328,8 +348,9 @@ def category(request):
 	return render(request, manageThemeDir + 'category.html', context)
 
 def category_add(request):
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
 	context = {}
-	
 	if request.method == 'POST':
 
 		chosenPid  = int(request.POST.get('pid', 0))
@@ -352,7 +373,8 @@ def category_add(request):
 
 
 def category_edit(request):
-	
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
 	context  = {}
 	cid      = int(request.GET.get('cid'))
 	detail   = Category.objects.get(category_id=cid)
@@ -370,6 +392,8 @@ def category_edit(request):
 
 
 def category_delete(request):
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
 	idList = request.REQUEST.getlist('idlist')
 	if len(idList) > 0:
 		Category.objects.filter(category_id__in=idList).delete()
@@ -381,6 +405,9 @@ def category_delete(request):
 
 
 def attachment(request):
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
+
 	context = {}
 	sql = 'SELECT username, attrch_id, original_name, create_time, size '
 	sql += 'FROM "' + Meta.db_table + '_attachment" AS attach LEFT JOIN "' + Meta.db_table +'_user" AS userx '
@@ -419,6 +446,9 @@ def attachment(request):
 
 
 def attachment_delete(request):
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
+
 	idList = request.REQUEST.getlist('idlist')
 	if len(idList) > 0:
 		Attachment.objects.filter(attrch_id__in=idList).delete()
@@ -430,6 +460,9 @@ def attachment_delete(request):
 
 
 def user(request):
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
+
 	context = {}
 	if request.GET.get('word'):
 		userList = User.objects.filter(username__icontains=cgi.escape(request.GET.get('word'))).order_by('-user_id')
@@ -454,6 +487,9 @@ def user(request):
 
 
 def user_add(request):
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
+
 	if request.method == 'POST':
 		email     = request.POST.get('email', '')
 		group     = cgi.escape(request.POST.get('group', ''))
@@ -475,6 +511,9 @@ def user_add(request):
 
 
 def user_edit(request):
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
+
     context  = {}
     uid      = int(request.GET.get('uid', 0))
 
@@ -495,6 +534,9 @@ def user_edit(request):
 
 
 def user_delete(request):
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
+
 	idList = request.REQUEST.getlist('idlist')
 	if len(idList) > 0:
 		User.objects.filter(user_id__in=idList).delete()
@@ -506,6 +548,9 @@ def user_delete(request):
 
 
 def themes(request):
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
+
 	context = {
 		'currTheme'   : C.getTheme(),
 		'configList'  : C.getThemesConfig()
@@ -515,6 +560,9 @@ def themes(request):
 
 
 def theme_swich(request):
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
+
     opt = Options.objects.get(name='theme')
     opt.value = cgi.escape(request.POST.get('change'))
     opt.save()
@@ -522,6 +570,9 @@ def theme_swich(request):
 
 
 def theme_edit(request):
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
+
     if request.GET.get('file'):
         fileName = cgi.escape(request.GET.get('file').replace('..', '').replace('/', '').replace('%', ''))
     else:
@@ -554,6 +605,9 @@ def theme_edit(request):
 
 
 def theme_changfile(request):
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
+
     if request.method == 'POST':
         themeDir = cgi.escape(request.POST.get('themeDir').replace('.', ''))
         f = open(os.path.split(os.path.realpath(__file__))[0] + '/templates/' + themeDir + cgi.escape(request.POST.get('file')), 'w')
@@ -566,9 +620,8 @@ def theme_changfile(request):
 
 
 def manage_profile(request):
-	userInfo = request.session.get('uInfo', False)
-	if C.checkLoginAdmin(userInfo) == False:
-		return HttpResponseRedirect('/')
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
 
 	msg = ''
 	if request.method == 'POST':
@@ -587,6 +640,9 @@ def manage_profile(request):
 
 
 def manage_option(request):
+    if C.checkLoginAdmin(request.session.get('uInfo', False)) == False:
+        return HttpResponseRedirect('/signin/')
+
     if request.method == 'POST':
         for key,val in request.POST.iteritems():
             if key != 'csrfmiddlewaretoken':
@@ -609,6 +665,4 @@ def manage_option(request):
     return render(request, manageThemeDir + 'option.html', context)
 
 def test(request):
-    a = C.getTest()
-    b = C.getTest()
-    return HttpResponse(str(id(a)) + '<br>' + str(id(b)))
+    return HttpResponse(request.session['uInfo']['uname'])
